requirements.txt
tensorflow>=2.16,<3.0
mlflow>=2.12,<3.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.7,<3.0
scikit-learn>=1.4,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
pyyaml>=6.0,<7.0
Pillow>=10.3,<12.0
python-multipart>=0.0.9
dvc>=3.49,<4.0
opencv-python-headless>=4.9,<5.0
